+++ src/test/java/walkingkooka/validation/form/provider/FilteredMappedFormHandlerProviderTest.java
@@ -38,7 +38,7 @@ import java.util.Objects;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public final class FilteredMappedFormHandlerProviderTest implements FormHandlerProviderTesting<FilteredMappedFormHandlerProvider>,
+public final class FilteredMappedFormHandlerProviderTest implements FormHandlerProviderTesting2<FilteredMappedFormHandlerProvider>,
     ToStringTesting<FilteredMappedFormHandlerProvider> {
 
     private final static AbsoluteUrl URL = Url.parseAbsolute("https://example.com/form-handler-123");
